% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adventures_dark_and_deep_terrain_encounters.R
\name{bloch_settlement}
\alias{bloch_settlement}
\title{Random table to know if there's a settlement on a new hex. It's
included in the random terrain function too.}
\usage{
bloch_settlement()
}
\description{
Random table to know if there's a settlement on a new hex. It's
included in the random terrain function too.
}
