% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adventures_dark_and_deep_terrain_encounters.R
\name{bloch_terrain}
\alias{bloch_terrain}
\title{Generate random terrain when coming from another.
You need to specify which terrain you're coming from.}
\usage{
bloch_terrain(from = bloch_hex())
}
\arguments{
\item{from}{A character. One of badlands, desert, forest, hills, mountains, plains, scrub or swamp.}
}
\description{
Generate random terrain when coming from another.
You need to specify which terrain you're coming from.
}
