% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{rl}
\alias{rl}
\title{Roll a nroll dice with nsides sides and nmod modifier}
\usage{
rl(nroll, nsides, nmod = 0)
}
\arguments{
\item{nroll}{An integer for number of dice to roll}

\item{nsides}{An integer for number of sides to of the dice}

\item{nmod}{An integer for a modifier to added to the result}
}
\value{
A string indicating all the dice rolled in xdy+z notation and the result
}
\description{
Roll a nroll dice with nsides sides and nmod modifier
}
